fix booking form
