+++ booking-app/app/media-commons/book/form/page.tsx
@@ -4,7 +4,10 @@
 
 import BookingFormDetailsPage from "@/components/src/client/routes/booking/formPages/BookingFormDetailsPage";
 import React from "react";
+import { Tenants } from "@/components/src/policy";
 
-const Form: React.FC = () => <BookingFormDetailsPage />;
+const Form: React.FC = () => (
+  <BookingFormDetailsPage tenant={Tenants.MEDIA_COMMONS} />
+);
 
 export default Form;
