+++ booking-app/app/media-commons/modification/form/[id]/page.tsx
@@ -1,15 +1,17 @@
-// app/modification/form/[id].tsx
+// app/media-commons/modification/form/[id].tsx
 
 "use client";
 
 import BookingFormDetailsPage from "@/components/src/client/routes/booking/formPages/BookingFormDetailsPage";
 import { FormContextLevel } from "@/components/src/types";
 import React from "react";
+import { Tenants } from "@/components/src/policy";
 
 const Form: React.FC = ({ params }: { params: { id: string } }) => (
   <BookingFormDetailsPage
     calendarEventId={params.id}
     formContext={FormContextLevel.MODIFICATION}
+    tenant={Tenants.MEDIA_COMMONS}
   />
 );
